refactor(agent): extract cost, history, and tool_execution modules from loop_
Split the 10K-line loop_.rs into focused modules:
- cost.rs: cost tracking context, budget checks, usage recording
- history.rs: history trimming, truncation, session state persistence
- tool_execution.rs: single/parallel/sequential tool execution

Also inlined 8 single-call wrapper functions and removed 6 dead-code
items (auto_compact_history, build_compaction_transcript, etc.).
