+++ src/agent/loop_.rs
@@ -1,8 +1,6 @@
 use crate::approval::{ApprovalManager, ApprovalRequest, ApprovalResponse};
-use crate::config::schema::ModelPricing;
 use crate::config::Config;
-use crate::cost::types::{BudgetCheck, TokenUsage as CostTokenUsage};
-use crate::cost::CostTracker;
+use crate::cost::types::BudgetCheck;
 use crate::i18n::ToolDescriptions;
 use crate::memory::{self, decay, Memory, MemoryCategory};
 use crate::multimodal;
@@ -18,117 +16,20 @@ use crate::util::truncate_with_ellipsis;
 use anyhow::Result;
 use futures_util::StreamExt;
 use regex::{Regex, RegexSet};
-use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
 use std::fmt::Write;
 use std::io::Write as _;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::{Arc, LazyLock, Mutex};
 use std::time::{Duration, Instant};
 use tokio_util::sync::CancellationToken;
 use uuid::Uuid;
 
-// ── Cost tracking via task-local ──
-
-/// Context for cost tracking within the tool call loop.
-/// Scoped via `tokio::task_local!` at call sites (channels, gateway).
-#[derive(Clone)]
-pub(crate) struct ToolLoopCostTrackingContext {
-    pub tracker: Arc<CostTracker>,
-    pub prices: Arc<std::collections::HashMap<String, ModelPricing>>,
-}
-
-impl ToolLoopCostTrackingContext {
-    pub(crate) fn new(
-        tracker: Arc<CostTracker>,
-        prices: Arc<std::collections::HashMap<String, ModelPricing>>,
-    ) -> Self {
-        Self { tracker, prices }
-    }
-}
-
-tokio::task_local! {
-    pub(crate) static TOOL_LOOP_COST_TRACKING_CONTEXT: Option<ToolLoopCostTrackingContext>;
-}
-
-/// 3-tier model pricing lookup:
-/// 1. Direct model name
-/// 2. Qualified `provider/model`
-/// 3. Suffix after last `/`
-fn lookup_model_pricing<'a>(
-    prices: &'a std::collections::HashMap<String, ModelPricing>,
-    provider_name: &str,
-    model: &str,
-) -> Option<&'a ModelPricing> {
-    prices
-        .get(model)
-        .or_else(|| prices.get(&format!("{provider_name}/{model}")))
-        .or_else(|| {
-            model
-                .rsplit_once('/')
-                .and_then(|(_, suffix)| prices.get(suffix))
-        })
-}
-
-/// Record token usage from an LLM response via the task-local cost tracker.
-/// Returns `(total_tokens, cost_usd)` on success, `None` when not scoped or no usage.
-fn record_tool_loop_cost_usage(
-    provider_name: &str,
-    model: &str,
-    usage: &crate::providers::traits::TokenUsage,
-) -> Option<(u64, f64)> {
-    let input_tokens = usage.input_tokens.unwrap_or(0);
-    let output_tokens = usage.output_tokens.unwrap_or(0);
-    let total_tokens = input_tokens.saturating_add(output_tokens);
-    if total_tokens == 0 {
-        return None;
-    }
-
-    let ctx = TOOL_LOOP_COST_TRACKING_CONTEXT
-        .try_with(Clone::clone)
-        .ok()
-        .flatten()?;
-    let pricing = lookup_model_pricing(&ctx.prices, provider_name, model);
-    let cost_usage = CostTokenUsage::new(
-        model,
-        input_tokens,
-        output_tokens,
-        pricing.map_or(0.0, |entry| entry.input),
-        pricing.map_or(0.0, |entry| entry.output),
-    );
-
-    if pricing.is_none() {
-        tracing::debug!(
-            provider = provider_name,
-            model,
-            "Cost tracking recorded token usage with zero pricing (no pricing entry found)"
-        );
-    }
-
-    if let Err(error) = ctx.tracker.record_usage(cost_usage.clone()) {
-        tracing::warn!(
-            provider = provider_name,
-            model,
-            "Failed to record cost tracking usage: {error}"
-        );
-    }
-
-    Some((cost_usage.total_tokens, cost_usage.cost_usd))
-}
-
-/// Check budget before an LLM call. Returns `None` when no cost tracking
-/// context is scoped (tests, delegate, CLI without cost config).
-pub(crate) fn check_tool_loop_budget() -> Option<BudgetCheck> {
-    TOOL_LOOP_COST_TRACKING_CONTEXT
-        .try_with(Clone::clone)
-        .ok()
-        .flatten()
-        .map(|ctx| {
-            ctx.tracker
-                .check_budget(0.0)
-                .unwrap_or(BudgetCheck::Allowed)
-        })
-}
+// Cost tracking moved to `super::cost`.
+pub(crate) use super::cost::{
+    check_tool_loop_budget, record_tool_loop_cost_usage, ToolLoopCostTrackingContext,
+    TOOL_LOOP_COST_TRACKING_CONTEXT,
+};
 
 /// Minimum characters per chunk when relaying LLM text to a streaming draft.
 const STREAM_CHUNK_MIN_CHARS: usize = 80;
@@ -139,95 +40,12 @@ const STREAM_TOOL_MARKER_WINDOW_CHARS: usize = 512;
 /// Used as a safe fallback when `max_tool_iterations` is unset or configured as zero.
 const DEFAULT_MAX_TOOL_ITERATIONS: usize = 10;
 
-/// Find the largest byte index `<= i` that is a valid char boundary.
-/// MSRV-compatible replacement for `str::floor_char_boundary` (stable in 1.91).
-fn floor_char_boundary(s: &str, i: usize) -> usize {
-    if i >= s.len() {
-        return s.len();
-    }
-    let mut pos = i;
-    while pos > 0 && !s.is_char_boundary(pos) {
-        pos -= 1;
-    }
-    pos
-}
-
-/// Find the smallest byte index `>= i` that is a valid char boundary.
-/// MSRV-compatible replacement for `str::ceil_char_boundary` (stable in 1.91).
-fn ceil_char_boundary(s: &str, i: usize) -> usize {
-    if i >= s.len() {
-        return s.len();
-    }
-    let mut pos = i;
-    while pos < s.len() && !s.is_char_boundary(pos) {
-        pos += 1;
-    }
-    pos
-}
-
-/// Truncate a tool result to `max_chars`, keeping head (2/3) + tail (1/3)
-/// with a marker in the middle. Returns input unchanged if within limit or
-/// `max_chars == 0` (disabled).
-pub(crate) fn truncate_tool_result(output: &str, max_chars: usize) -> String {
-    if max_chars == 0 || output.len() <= max_chars {
-        return output.to_string();
-    }
-    let head_len = max_chars * 2 / 3;
-    let tail_len = max_chars.saturating_sub(head_len);
-    let head_end = floor_char_boundary(output, head_len);
-    let tail_start = ceil_char_boundary(output, output.len().saturating_sub(tail_len));
-    // Guard against overlap when max_chars is very small
-    if head_end >= tail_start {
-        return output[..floor_char_boundary(output, max_chars)].to_string();
-    }
-    let truncated_chars = tail_start - head_end;
-    format!(
-        "{}\n\n[... {} characters truncated ...]\n\n{}",
-        &output[..head_end],
-        truncated_chars,
-        &output[tail_start..]
-    )
-}
-
-/// Aggressively trim old tool result messages in history to recover from
-/// context overflow. Keeps the last `protect_last_n` messages untouched.
-/// Returns total characters saved.
-fn fast_trim_tool_results(
-    history: &mut [crate::providers::ChatMessage],
-    protect_last_n: usize,
-) -> usize {
-    let trim_to = 2000;
-    let mut saved = 0;
-    let cutoff = history.len().saturating_sub(protect_last_n);
-    for msg in &mut history[..cutoff] {
-        if msg.role == "tool" && msg.content.len() > trim_to {
-            let original_len = msg.content.len();
-            msg.content = truncate_tool_result(&msg.content, trim_to);
-            saved += original_len - msg.content.len();
-        }
-    }
-    saved
-}
-
-/// Emergency: drop oldest non-system, non-recent messages from history.
-/// Returns number of messages dropped.
-fn emergency_history_trim(
-    history: &mut Vec<crate::providers::ChatMessage>,
-    keep_recent: usize,
-) -> usize {
-    let mut dropped = 0;
-    let target_drop = history.len() / 3;
-    let mut i = 0;
-    while dropped < target_drop && i < history.len().saturating_sub(keep_recent) {
-        if history[i].role == "system" {
-            i += 1;
-        } else {
-            history.remove(i);
-            dropped += 1;
-        }
-    }
-    dropped
-}
+// History management moved to `super::history`.
+pub(crate) use super::history::{
+    emergency_history_trim, estimate_history_tokens, fast_trim_tool_results,
+    load_interactive_session_history, save_interactive_session_history, trim_history,
+    truncate_tool_result,
+};
 
 /// Minimum user-message length (in chars) for auto-save to memory.
 /// Matches the channel-side constant in `channels/mod.rs`.
@@ -425,29 +243,6 @@ pub(crate) fn scrub_credentials(input: &str) -> String {
 /// Default trigger for auto-compaction when non-system message count exceeds this threshold.
 /// Prefer passing the config-driven value via `run_tool_call_loop`; this constant is only
 /// used when callers omit the parameter.
-const DEFAULT_MAX_HISTORY_MESSAGES: usize = 50;
-
-/// Keep this many most-recent non-system messages after compaction.
-const COMPACTION_KEEP_RECENT_MESSAGES: usize = 20;
-
-/// Safety cap for compaction source transcript passed to the summarizer.
-const COMPACTION_MAX_SOURCE_CHARS: usize = 12_000;
-
-/// Max characters retained in stored compaction summary.
-const COMPACTION_MAX_SUMMARY_CHARS: usize = 2_000;
-
-/// Estimate token count for a message history using ~4 chars/token heuristic.
-/// Includes a small overhead per message for role/framing tokens.
-fn estimate_history_tokens(history: &[ChatMessage]) -> usize {
-    history
-        .iter()
-        .map(|m| {
-            // ~4 chars per token + ~4 framing tokens per message (role, delimiters)
-            m.content.len().div_ceil(4) + 4
-        })
-        .sum()
-}
-
 /// Minimum interval between progress sends to avoid flooding the draft channel.
 pub(crate) const PROGRESS_MIN_INTERVAL_MS: u64 = 500;
 
@@ -466,22 +261,6 @@ pub enum DraftEvent {
 
 tokio::task_local! {
     pub(crate) static TOOL_CHOICE_OVERRIDE: Option<String>;
-}
-
-/// Extract a short hint from tool call arguments for progress display.
-fn truncate_tool_args_for_progress(name: &str, args: &serde_json::Value, max_len: usize) -> String {
-    let hint = match name {
-        "shell" => args.get("command").and_then(|v| v.as_str()),
-        "file_read" | "file_write" => args.get("path").and_then(|v| v.as_str()),
-        _ => args
-            .get("action")
-            .and_then(|v| v.as_str())
-            .or_else(|| args.get("query").and_then(|v| v.as_str())),
-    };
-    match hint {
-        Some(s) => truncate_with_ellipsis(s, max_len),
-        None => String::new(),
-    }
 }
 
 /// Convert a tool registry to OpenAI function-calling format for native tool support.
@@ -503,162 +282,6 @@ fn tools_to_openai_format(tools_registry: &[Box<dyn Tool>]) -> Vec<serde_json::V
 
 fn autosave_memory_key(prefix: &str) -> String {
     format!("{prefix}_{}", Uuid::new_v4())
-}
-
-fn memory_session_id_from_state_file(path: &Path) -> Option<String> {
-    let raw = path.to_string_lossy().trim().to_string();
-    if raw.is_empty() {
-        return None;
-    }
-
-    Some(format!("cli:{raw}"))
-}
-
-/// Trim conversation history to prevent unbounded growth.
-/// Preserves the system prompt (first message if role=system) and the most recent messages.
-fn trim_history(history: &mut Vec<ChatMessage>, max_history: usize) {
-    // Nothing to trim if within limit
-    let has_system = history.first().map_or(false, |m| m.role == "system");
-    let non_system_count = if has_system {
-        history.len() - 1
-    } else {
-        history.len()
-    };
-
-    if non_system_count <= max_history {
-        return;
-    }
-
-    let start = if has_system { 1 } else { 0 };
-    let to_remove = non_system_count - max_history;
-    history.drain(start..start + to_remove);
-}
-
-fn build_compaction_transcript(messages: &[ChatMessage]) -> String {
-    let mut transcript = String::new();
-    for msg in messages {
-        let role = msg.role.to_uppercase();
-        let _ = writeln!(transcript, "{role}: {}", msg.content.trim());
-    }
-
-    if transcript.chars().count() > COMPACTION_MAX_SOURCE_CHARS {
-        truncate_with_ellipsis(&transcript, COMPACTION_MAX_SOURCE_CHARS)
-    } else {
-        transcript
-    }
-}
-
-fn apply_compaction_summary(
-    history: &mut Vec<ChatMessage>,
-    start: usize,
-    compact_end: usize,
-    summary: &str,
-) {
-    let summary_msg = ChatMessage::assistant(format!("[Compaction summary]\n{}", summary.trim()));
-    history.splice(start..compact_end, std::iter::once(summary_msg));
-}
-
-async fn auto_compact_history(
-    history: &mut Vec<ChatMessage>,
-    provider: &dyn Provider,
-    model: &str,
-    max_history: usize,
-    max_context_tokens: usize,
-) -> Result<bool> {
-    let has_system = history.first().map_or(false, |m| m.role == "system");
-    let non_system_count = if has_system {
-        history.len().saturating_sub(1)
-    } else {
-        history.len()
-    };
-
-    let estimated_tokens = estimate_history_tokens(history);
-
-    // Trigger compaction when either token budget OR message count is exceeded.
-    if estimated_tokens <= max_context_tokens && non_system_count <= max_history {
-        return Ok(false);
-    }
-
-    let start = if has_system { 1 } else { 0 };
-    let keep_recent = COMPACTION_KEEP_RECENT_MESSAGES.min(non_system_count);
-    let compact_count = non_system_count.saturating_sub(keep_recent);
-    if compact_count == 0 {
-        return Ok(false);
-    }
-
-    let mut compact_end = start + compact_count;
-
-    // Snap compact_end to a user-turn boundary so we don't split mid-conversation.
-    while compact_end > start && history.get(compact_end).map_or(false, |m| m.role != "user") {
-        compact_end -= 1;
-    }
-    if compact_end <= start {
-        return Ok(false);
-    }
-
-    let to_compact: Vec<ChatMessage> = history[start..compact_end].to_vec();
-    let transcript = build_compaction_transcript(&to_compact);
-
-    let summarizer_system = "You are a conversation compaction engine. Summarize older chat history into concise context for future turns. Preserve: user preferences, commitments, decisions, unresolved tasks, key facts. Omit: filler, repeated chit-chat, verbose tool logs. Output plain text bullet points only.";
-
-    let summarizer_user = format!(
-        "Summarize the following conversation history for context preservation. Keep it short (max 12 bullet points).\n\n{}",
-        transcript
-    );
-
-    let summary_raw = provider
-        .chat_with_system(Some(summarizer_system), &summarizer_user, model, 0.2)
-        .await
-        .unwrap_or_else(|_| {
-            // Fallback to deterministic local truncation when summarization fails.
-            truncate_with_ellipsis(&transcript, COMPACTION_MAX_SUMMARY_CHARS)
-        });
-
-    let summary = truncate_with_ellipsis(&summary_raw, COMPACTION_MAX_SUMMARY_CHARS);
-    apply_compaction_summary(history, start, compact_end, &summary);
-
-    Ok(true)
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct InteractiveSessionState {
-    version: u32,
-    history: Vec<ChatMessage>,
-}
-
-impl InteractiveSessionState {
-    fn from_history(history: &[ChatMessage]) -> Self {
-        Self {
-            version: 1,
-            history: history.to_vec(),
-        }
-    }
-}
-
-fn load_interactive_session_history(path: &Path, system_prompt: &str) -> Result<Vec<ChatMessage>> {
-    if !path.exists() {
-        return Ok(vec![ChatMessage::system(system_prompt)]);
-    }
-
-    let raw = std::fs::read_to_string(path)?;
-    let mut state: InteractiveSessionState = serde_json::from_str(&raw)?;
-    if state.history.is_empty() {
-        state.history.push(ChatMessage::system(system_prompt));
-    } else if state.history.first().map(|msg| msg.role.as_str()) != Some("system") {
-        state.history.insert(0, ChatMessage::system(system_prompt));
-    }
-
-    Ok(state.history)
-}
-
-fn save_interactive_session_history(path: &Path, history: &[ChatMessage]) -> Result<()> {
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-
-    let payload = serde_json::to_string_pretty(&InteractiveSessionState::from_history(history))?;
-    std::fs::write(path, payload)?;
-    Ok(())
 }
 
 /// Build context preamble by searching memory for relevant entries.
@@ -754,10 +377,11 @@ fn build_hardware_context(
     context
 }
 
-/// Find a tool by name in the registry.
-fn find_tool<'a>(tools: &'a [Box<dyn Tool>], name: &str) -> Option<&'a dyn Tool> {
-    tools.iter().find(|t| t.name() == name).map(|t| t.as_ref())
-}
+// Tool execution moved to `super::tool_execution`.
+pub(crate) use super::tool_execution::{
+    execute_tools_parallel, execute_tools_sequential, should_execute_tools_in_parallel,
+    ToolExecutionOutcome,
+};
 
 fn parse_arguments_value(raw: Option<&serde_json::Value>) -> serde_json::Value {
     match raw {
@@ -804,12 +428,6 @@ fn canonicalize_json_for_tool_signature(value: &serde_json::Value) -> serde_json
         ),
         _ => value.clone(),
     }
-}
-
-fn tool_call_signature(name: &str, arguments: &serde_json::Value) -> (String, String) {
-    let canonical_args = canonicalize_json_for_tool_signature(arguments);
-    let args_json = serde_json::to_string(&canonical_args).unwrap_or_else(|_| "{}".to_string());
-    (name.trim().to_ascii_lowercase(), args_json)
 }
 
 fn parse_tool_call_value(value: &serde_json::Value) -> Option<ParsedToolCall> {
@@ -1128,18 +746,6 @@ fn find_first_tag<'a>(haystack: &str, tags: &'a [&'a str]) -> Option<(usize, &'a
     tags.iter()
         .filter_map(|tag| haystack.find(tag).map(|idx| (idx, *tag)))
         .min_by_key(|(idx, _)| *idx)
-}
-
-fn matching_tool_call_close_tag(open_tag: &str) -> Option<&'static str> {
-    match open_tag {
-        "<tool_call>" => Some("</tool_call>"),
-        "<toolcall>" => Some("</toolcall>"),
-        "<tool-call>" => Some("</tool-call>"),
-        "<invoke>" => Some("</invoke>"),
-        "<minimax:tool_call>" => Some("</minimax:tool_call>"),
-        "<minimax:toolcall>" => Some("</minimax:toolcall>"),
-        _ => None,
-    }
 }
 
 fn extract_first_json_value_with_end(input: &str) -> Option<(serde_json::Value, usize)> {
@@ -1795,7 +1401,15 @@ fn parse_tool_calls(response: &str) -> (String, Vec<ParsedToolCall>) {
             text_parts.push(before.trim().to_string());
         }
 
-        let Some(close_tag) = matching_tool_call_close_tag(open_tag) else {
+        let Some(close_tag) = (match open_tag {
+            "<tool_call>" => Some("</tool_call>"),
+            "<toolcall>" => Some("</toolcall>"),
+            "<tool-call>" => Some("</tool-call>"),
+            "<invoke>" => Some("</invoke>"),
+            "<minimax:tool_call>" => Some("</minimax:tool_call>"),
+            "<minimax:toolcall>" => Some("</minimax:toolcall>"),
+            _ => None,
+        }) else {
             break;
         };
 
@@ -2224,18 +1838,6 @@ fn detect_tool_call_parse_issue(response: &str, parsed_calls: &[ParsedToolCall])
     }
 }
 
-fn parse_structured_tool_calls(tool_calls: &[ToolCall]) -> Vec<ParsedToolCall> {
-    tool_calls
-        .iter()
-        .map(|call| ParsedToolCall {
-            name: call.name.clone(),
-            arguments: serde_json::from_str::<serde_json::Value>(&call.arguments)
-                .unwrap_or_else(|_| serde_json::Value::Object(serde_json::Map::new())),
-            tool_call_id: Some(call.id.clone()),
-        })
-        .collect()
-}
-
 /// Build assistant history entry in JSON format for native tool-call APIs.
 /// `convert_messages` in the OpenRouter provider parses this JSON to reconstruct
 /// the proper `NativeMessage` with structured `tool_calls`.
@@ -2313,27 +1915,6 @@ fn build_native_assistant_history_from_parsed_calls(
     Some(obj.to_string())
 }
 
-fn build_assistant_history_with_tool_calls(text: &str, tool_calls: &[ToolCall]) -> String {
-    let mut parts = Vec::new();
-
-    if !text.trim().is_empty() {
-        parts.push(text.trim().to_string());
-    }
-
-    for call in tool_calls {
-        let arguments = serde_json::from_str::<serde_json::Value>(&call.arguments)
-            .unwrap_or_else(|_| serde_json::Value::String(call.arguments.clone()));
-        let payload = serde_json::json!({
-            "id": call.id,
-            "name": call.name,
-            "arguments": arguments,
-        });
-        parts.push(format!("<tool_call>\n{payload}\n</tool_call>"));
-    }
-
-    parts.join("\n")
-}
-
 fn resolve_display_text(
     response_text: &str,
     parsed_text: &str,
@@ -2358,10 +1939,10 @@ fn resolve_display_text(
 }
 
 #[derive(Debug, Clone)]
-struct ParsedToolCall {
-    name: String,
-    arguments: serde_json::Value,
-    tool_call_id: Option<String>,
+pub(crate) struct ParsedToolCall {
+    pub(crate) name: String,
+    pub(crate) arguments: serde_json::Value,
+    pub(crate) tool_call_id: Option<String>,
 }
 
 #[derive(Debug)]
@@ -2411,39 +1992,6 @@ struct StreamedChatOutcome {
     forwarded_live_deltas: bool,
 }
 
-fn looks_like_streamed_tool_payload(window: &str) -> bool {
-    let lowered = window.to_ascii_lowercase();
-    lowered.contains("<tool_call")
-        || lowered.contains("<toolcall")
-        || lowered.contains("\"tool_calls\"")
-}
-
-async fn call_provider_chat(
-    provider: &dyn Provider,
-    messages: &[ChatMessage],
-    request_tools: Option<&[crate::tools::ToolSpec]>,
-    model: &str,
-    temperature: f64,
-    cancellation_token: Option<&CancellationToken>,
-) -> Result<crate::providers::ChatResponse> {
-    let chat_future = provider.chat(
-        ChatRequest {
-            messages,
-            tools: request_tools,
-        },
-        model,
-        temperature,
-    );
-
-    if let Some(token) = cancellation_token {
-        tokio::select! {
-            () = token.cancelled() => Err(ToolLoopCancelled.into()),
-            result = chat_future => result,
-        }
-    } else {
-        chat_future.await
-    }
-}
 
 async fn consume_provider_streaming_response(
     provider: &dyn Provider,
@@ -2517,7 +2065,12 @@ async fn consume_provider_streaming_response(
                     marker_window.drain(..boundary);
                 }
 
-                if !suppress_forwarding && looks_like_streamed_tool_payload(&marker_window) {
+                if !suppress_forwarding && {
+                    let lowered = marker_window.to_ascii_lowercase();
+                    lowered.contains("<tool_call")
+                        || lowered.contains("<toolcall")
+                        || lowered.contains("\"tool_calls\"")
+                } {
                     suppress_forwarding = true;
                     if outcome.forwarded_live_deltas {
                         if let Some(tx) = delta_sender {
@@ -2691,181 +2244,6 @@ fn maybe_inject_channel_delivery_defaults(
         }
         Some(_) => {}
     }
-}
-
-async fn execute_one_tool(
-    call_name: &str,
-    call_arguments: serde_json::Value,
-    tools_registry: &[Box<dyn Tool>],
-    activated_tools: Option<&std::sync::Arc<std::sync::Mutex<crate::tools::ActivatedToolSet>>>,
-    observer: &dyn Observer,
-    cancellation_token: Option<&CancellationToken>,
-) -> Result<ToolExecutionOutcome> {
-    let args_summary = truncate_with_ellipsis(&call_arguments.to_string(), 300);
-    observer.record_event(&ObserverEvent::ToolCallStart {
-        tool: call_name.to_string(),
-        arguments: Some(args_summary),
-    });
-    let start = Instant::now();
-
-    let static_tool = find_tool(tools_registry, call_name);
-    let activated_arc = if static_tool.is_none() {
-        activated_tools.and_then(|at| at.lock().unwrap().get_resolved(call_name))
-    } else {
-        None
-    };
-    let Some(tool) = static_tool.or(activated_arc.as_deref()) else {
-        let reason = format!("Unknown tool: {call_name}");
-        let duration = start.elapsed();
-        observer.record_event(&ObserverEvent::ToolCall {
-            tool: call_name.to_string(),
-            duration,
-            success: false,
-        });
-        return Ok(ToolExecutionOutcome {
-            output: reason.clone(),
-            success: false,
-            error_reason: Some(scrub_credentials(&reason)),
-            duration,
-        });
-    };
-
-    let tool_future = tool.execute(call_arguments);
-    let tool_result = if let Some(token) = cancellation_token {
-        tokio::select! {
-            () = token.cancelled() => return Err(ToolLoopCancelled.into()),
-            result = tool_future => result,
-        }
-    } else {
-        tool_future.await
-    };
-
-    match tool_result {
-        Ok(r) => {
-            let duration = start.elapsed();
-            observer.record_event(&ObserverEvent::ToolCall {
-                tool: call_name.to_string(),
-                duration,
-                success: r.success,
-            });
-            if r.success {
-                Ok(ToolExecutionOutcome {
-                    output: scrub_credentials(&r.output),
-                    success: true,
-                    error_reason: None,
-                    duration,
-                })
-            } else {
-                let reason = r.error.unwrap_or(r.output);
-                Ok(ToolExecutionOutcome {
-                    output: format!("Error: {reason}"),
-                    success: false,
-                    error_reason: Some(scrub_credentials(&reason)),
-                    duration,
-                })
-            }
-        }
-        Err(e) => {
-            let duration = start.elapsed();
-            observer.record_event(&ObserverEvent::ToolCall {
-                tool: call_name.to_string(),
-                duration,
-                success: false,
-            });
-            let reason = format!("Error executing {call_name}: {e}");
-            Ok(ToolExecutionOutcome {
-                output: reason.clone(),
-                success: false,
-                error_reason: Some(scrub_credentials(&reason)),
-                duration,
-            })
-        }
-    }
-}
-
-struct ToolExecutionOutcome {
-    output: String,
-    success: bool,
-    error_reason: Option<String>,
-    duration: Duration,
-}
-
-fn should_execute_tools_in_parallel(
-    tool_calls: &[ParsedToolCall],
-    approval: Option<&ApprovalManager>,
-) -> bool {
-    if tool_calls.len() <= 1 {
-        return false;
-    }
-
-    // tool_search activates deferred MCP tools into ActivatedToolSet.
-    // Running tool_search in parallel with the tools it activates causes a
-    // race condition where the tool lookup happens before activation completes.
-    // Force sequential execution whenever tool_search is in the batch.
-    if tool_calls.iter().any(|call| call.name == "tool_search") {
-        return false;
-    }
-
-    if let Some(mgr) = approval {
-        if tool_calls.iter().any(|call| mgr.needs_approval(&call.name)) {
-            // Approval-gated calls must keep sequential handling so the caller can
-            // enforce CLI prompt/deny policy consistently.
-            return false;
-        }
-    }
-
-    true
-}
-
-async fn execute_tools_parallel(
-    tool_calls: &[ParsedToolCall],
-    tools_registry: &[Box<dyn Tool>],
-    activated_tools: Option<&std::sync::Arc<std::sync::Mutex<crate::tools::ActivatedToolSet>>>,
-    observer: &dyn Observer,
-    cancellation_token: Option<&CancellationToken>,
-) -> Result<Vec<ToolExecutionOutcome>> {
-    let futures: Vec<_> = tool_calls
-        .iter()
-        .map(|call| {
-            execute_one_tool(
-                &call.name,
-                call.arguments.clone(),
-                tools_registry,
-                activated_tools,
-                observer,
-                cancellation_token,
-            )
-        })
-        .collect();
-
-    let results = futures_util::future::join_all(futures).await;
-    results.into_iter().collect()
-}
-
-async fn execute_tools_sequential(
-    tool_calls: &[ParsedToolCall],
-    tools_registry: &[Box<dyn Tool>],
-    activated_tools: Option<&std::sync::Arc<std::sync::Mutex<crate::tools::ActivatedToolSet>>>,
-    observer: &dyn Observer,
-    cancellation_token: Option<&CancellationToken>,
-) -> Result<Vec<ToolExecutionOutcome>> {
-    let mut outcomes = Vec::with_capacity(tool_calls.len());
-
-    for call in tool_calls {
-        outcomes.push(
-            execute_one_tool(
-                &call.name,
-                call.arguments.clone(),
-                tools_registry,
-                activated_tools,
-                observer,
-                cancellation_token,
-            )
-            .await?,
-        );
-    }
-
-    Ok(outcomes)
 }
 
 // ── Agent Tool-Call Loop ──────────────────────────────────────────────────
@@ -3191,15 +2569,24 @@ pub(crate) async fn run_tool_call_loop(
                     if let Some(ref tx) = on_delta {
                         let _ = tx.send(DraftEvent::Clear).await;
                     }
-                    call_provider_chat(
-                        active_provider,
-                        &prepared_messages.messages,
-                        request_tools,
-                        active_model,
-                        temperature,
-                        cancellation_token.as_ref(),
-                    )
-                    .await
+                    {
+                        let chat_future = active_provider.chat(
+                            ChatRequest {
+                                messages: &prepared_messages.messages,
+                                tools: request_tools,
+                            },
+                            active_model,
+                            temperature,
+                        );
+                        if let Some(token) = cancellation_token.as_ref() {
+                            tokio::select! {
+                                () = token.cancelled() => Err(ToolLoopCancelled.into()),
+                                result = chat_future => result,
+                            }
+                        } else {
+                            chat_future.await
+                        }
+                    }
                 }
             }
         } else {
@@ -3288,7 +2675,16 @@ pub(crate) async fn run_tool_call_loop(
                 // Fall back to text-based parsing (XML tags, markdown blocks,
                 // GLM format) only if the provider returned no native calls —
                 // this ensures we support both native and prompt-guided models.
-                let mut calls = parse_structured_tool_calls(&resp.tool_calls);
+                let mut calls: Vec<ParsedToolCall> = resp
+                    .tool_calls
+                    .iter()
+                    .map(|call| ParsedToolCall {
+                        name: call.name.clone(),
+                        arguments: serde_json::from_str::<serde_json::Value>(&call.arguments)
+                            .unwrap_or_else(|_| serde_json::Value::Object(serde_json::Map::new())),
+                        tool_call_id: Some(call.id.clone()),
+                    })
+                    .collect();
                 let mut parsed_text = String::new();
 
                 if calls.is_empty() {
@@ -3651,7 +3047,11 @@ pub(crate) async fn run_tool_call_loop(
                 }
             }
 
-            let signature = tool_call_signature(&tool_name, &tool_args);
+            let signature = {
+                let canonical_args = canonicalize_json_for_tool_signature(&tool_args);
+                let args_json = serde_json::to_string(&canonical_args).unwrap_or_else(|_| "{}".to_string());
+                (tool_name.trim().to_ascii_lowercase(), args_json)
+            };
             let dedup_exempt = dedup_exempt_tools.iter().any(|e| e == &tool_name);
             if !dedup_exempt && !seen_tool_signatures.insert(signature) {
                 let duplicate = format!(
@@ -3710,7 +3110,20 @@ pub(crate) async fn run_tool_call_loop(
 
             // ── Progress: tool start ────────────────────────────
             if let Some(ref tx) = on_delta {
-                let hint = truncate_tool_args_for_progress(&tool_name, &tool_args, 60);
+                let hint = {
+                    let raw = match tool_name.as_str() {
+                        "shell" => tool_args.get("command").and_then(|v| v.as_str()),
+                        "file_read" | "file_write" => tool_args.get("path").and_then(|v| v.as_str()),
+                        _ => tool_args
+                            .get("action")
+                            .and_then(|v| v.as_str())
+                            .or_else(|| tool_args.get("query").and_then(|v| v.as_str())),
+                    };
+                    match raw {
+                        Some(s) => truncate_with_ellipsis(s, 60),
+                        None => String::new(),
+                    }
+                };
                 let progress = if hint.is_empty() {
                     format!("\u{23f3} {}\n", tool_name)
                 } else {
@@ -4423,9 +3836,14 @@ pub async fn run(
         None
     };
     let channel_name = if interactive { "cli" } else { "daemon" };
-    let memory_session_id = session_state_file
-        .as_deref()
-        .and_then(memory_session_id_from_state_file);
+    let memory_session_id = session_state_file.as_deref().and_then(|path| {
+        let raw = path.to_string_lossy().trim().to_string();
+        if raw.is_empty() {
+            None
+        } else {
+            Some(format!("cli:{raw}"))
+        }
+    });
 
     // ── Execute ──────────────────────────────────────────────────
     let start = Instant::now();
@@ -5361,10 +4779,11 @@ pub async fn process_message(
 #[cfg(test)]
 mod tests {
     use super::{
-        apply_compaction_summary, build_compaction_transcript, emergency_history_trim,
-        estimate_history_tokens, fast_trim_tool_results, load_interactive_session_history,
-        save_interactive_session_history, truncate_tool_result, InteractiveSessionState,
+        emergency_history_trim, estimate_history_tokens, fast_trim_tool_results,
+        load_interactive_session_history, save_interactive_session_history, truncate_tool_result,
     };
+    use crate::agent::history::{InteractiveSessionState, DEFAULT_MAX_HISTORY_MESSAGES};
+    use crate::agent::tool_execution::execute_one_tool;
     use crate::providers::ChatMessage;
     use tempfile::tempdir;
 
@@ -8572,35 +7991,6 @@ Tail"#;
         ];
         trim_history(&mut history, DEFAULT_MAX_HISTORY_MESSAGES);
         assert_eq!(history.len(), 3);
-    }
-
-    #[test]
-    fn build_compaction_transcript_formats_roles() {
-        let messages = vec![
-            ChatMessage::user("I like dark mode"),
-            ChatMessage::assistant("Got it"),
-        ];
-        let transcript = build_compaction_transcript(&messages);
-        assert!(transcript.contains("USER: I like dark mode"));
-        assert!(transcript.contains("ASSISTANT: Got it"));
-    }
-
-    #[test]
-    fn apply_compaction_summary_replaces_old_segment() {
-        let mut history = vec![
-            ChatMessage::system("sys"),
-            ChatMessage::user("old 1"),
-            ChatMessage::assistant("old 2"),
-            ChatMessage::user("recent 1"),
-            ChatMessage::assistant("recent 2"),
-        ];
-
-        apply_compaction_summary(&mut history, 1, 3, "- user prefers concise replies");
-
-        assert_eq!(history.len(), 4);
-        assert!(history[1].content.contains("Compaction summary"));
-        assert!(history[2].content.contains("recent 1"));
-        assert!(history[3].content.contains("recent 2"));
     }
 
     #[test]
