+++ src/agent/mod.rs
@@ -3,11 +3,14 @@ pub mod agent;
 pub mod classifier;
 pub mod context_analyzer;
 pub mod context_compressor;
+pub mod cost;
 pub mod dispatcher;
 pub mod eval;
+pub mod history;
 pub mod history_pruner;
 pub mod loop_;
 pub mod loop_detector;
+pub mod tool_execution;
 pub mod memory_loader;
 pub mod personality;
 pub mod prompt;
